Moved optional macros
