+++ optional_configuration/ceasefire.hpp
