+++ optional_configuration/debug.hpp
