+++ optional_configuration/objectScaling.hpp
