+++ optional_configuration/sogConfig.hpp
