+++ startup/configuration/groups/serverConfigGroup.sqf
@@ -2,7 +2,7 @@
 // Includes all config scripts needed for the server.
 
 // AI Caching settings
-#include "..\..\..\configuration\aiCache.sqf"
+#include "..\..\..\optional_configuration\aiCache.sqf"
 
 // Lootbox
 #include "..\internals\lootBox.sqf"
