+++ startup/configuration/internals/configMacros.hpp
@@ -4,16 +4,16 @@
 // To disable a feature, comment out its #define line.
 
 #include "configAliases.hpp"
-#include "..\..\..\configuration\debug.hpp"
-#include "..\..\..\configuration\aceActions.hpp"
-#include "..\..\..\configuration\ceasefire.hpp"
-#include "..\..\..\configuration\killTracking.hpp"
+#include "..\..\..\optional_configuration\debug.hpp"
+#include "..\..\..\optional_configuration\aceActions.hpp"
+#include "..\..\..\optional_configuration\ceasefire.hpp"
+#include "..\..\..\optional_configuration\killTracking.hpp"
 #include "..\..\..\optional_configuration\gravestones.hpp"
-#include "..\..\..\configuration\viewDistanceEditor.hpp"
+#include "..\..\..\optional_configuration\viewDistanceEditor.hpp"
 #include "..\..\..\core_configuration\identityReplacement.hpp"
 #include "..\..\..\optional_configuration\insignia.hpp"
-#include "..\..\..\configuration\objectScaling.hpp"
-#include "..\..\..\configuration\sogConfig.hpp"
+#include "..\..\..\optional_configuration\objectScaling.hpp"
+#include "..\..\..\optional_configuration\sogConfig.hpp"
 #include "..\..\..\configuration\respawn.hpp"
 #include "..\..\..\configuration\slottingGenerator.hpp"
-#include "..\..\..\configuration\statsTracking.hpp"
+#include "..\..\..\optional_configuration\statsTracking.hpp"
